refactor: stock collection pagination 메소드 분리
